+++ src/core_modules/capture-core/components/Pages/NewRelationship/RegisterTei/RegisterTei.component.js
@@ -1,14 +1,16 @@
 // @flow
 import React, { type ComponentType, useContext, useState } from 'react';
+import { compose } from 'redux';
 import { withStyles } from '@material-ui/core/styles';
 import i18n from '@dhis2/d2-i18n';
 import { Button } from '../../../Buttons';
 import { RegisterTeiDataEntry } from './DataEntry/RegisterTeiDataEntry.container';
 import { RegistrationSection } from './RegistrationSection';
-import GeneralOutput from './GeneralOutput/GeneralOutput.container';
-import { ReviewDialog } from './GeneralOutput/WarningsSection/SearchGroupDuplicate/ReviewDialog.component';
+import { DataEntryWidgetOutput } from '../../../DataEntryWidgetOutput/DataEntryWidgetOutput.container';
+import { PossibleDuplicatesDialog } from '../../../PossibleDuplicatesDialog';
 import { ResultsPageSizeContext } from '../../shared-contexts';
 import type { Props } from './RegisterTei.types';
+import { withErrorMessageHandler, withLoadingIndicator } from '../../../../HOC';
 
 const getStyles = () => ({
     container: {
@@ -27,8 +29,10 @@ const RegisterTeiPlain = ({
     onSave,
     onReviewDuplicates,
     onGetUnsavedAttributeValues,
-    possibleDuplicates,
-    tetName,
+    dataEntryId,
+    possibleDuplicatesExist,
+    trackedEntityName,
+    newRelationshipProgramId,
     classes,
 }: Props) => {
     const { resultsPageSize } = useContext(ResultsPageSizeContext);
@@ -37,7 +41,7 @@ const RegisterTeiPlain = ({
     const [savedArguments, setArguments] = useState([]);
 
     function handleSaveAttempt(...args) {
-        if (possibleDuplicates) {
+        if (possibleDuplicatesExist) {
             setArguments(args);
             onReviewDuplicates(resultsPageSize);
             toggleDuplicatesModal(true);
@@ -60,7 +64,7 @@ const RegisterTeiPlain = ({
                     onClick={handleSaveFromDialog}
                     primary
                 >
-                    {i18n.t('Save as new {{tetName}}', { tetName })}
+                    {i18n.t('Save as new {{trackedEntityName}}', { trackedEntityName })}
                 </Button>
             </div>
         </React.Fragment>
@@ -84,10 +88,14 @@ const RegisterTeiPlain = ({
                     onGetUnsavedAttributeValues={onGetUnsavedAttributeValues}
                 />
             </div>
-            <GeneralOutput
+            <DataEntryWidgetOutput
+                dataEntryId={dataEntryId}
+                selectedScopeId={newRelationshipProgramId}
                 onLink={onLink}
             />
-            <ReviewDialog
+            <PossibleDuplicatesDialog
+                dataEntryId={dataEntryId}
+                selectedScopeId={newRelationshipProgramId}
                 open={duplicatesOpen}
                 onLink={onLink}
                 onCancel={handleDialogCancel}
@@ -97,4 +105,9 @@ const RegisterTeiPlain = ({
     );
 };
 
-export const RegisterTeiComponent: ComponentType<$Diff<Props, CssClasses>> = withStyles(getStyles)(RegisterTeiPlain);
+export const RegisterTeiComponent: ComponentType<$Diff<Props, CssClasses>> =
+  compose(
+      withLoadingIndicator(),
+      withErrorMessageHandler(),
+      withStyles(getStyles),
+  )(RegisterTeiPlain);