+++ src/core_modules/capture-core/components/PossibleDuplicatesDialog/ReviewDialogContents/ReviewDialogContentsPager.types.js
@@ -1,0 +1,15 @@
+// @flow
+
+export type OwnProps = {|
+    nextPageButtonDisabled: boolean,
+    dataEntryId: string,
+    selectedScopeId: string
+|};
+type DispatchersFromFromRedux = {|
+    onChangePage: Function,
+|};
+type PropsFromRedux = {|
+    currentPage: number,
+|};
+
+export type Props ={| ...OwnProps, ...DispatchersFromFromRedux, ...PropsFromRedux, ...CssClasses |}