+++ src/core_modules/capture-core/metaData/helpers/index.js
@@ -17,3 +17,4 @@ export {
 export { convertValues as convertDataElementsValues } from './DataElements';
 export { getScopeFromScopeId } from './getScopeFromScopeId';
 export { programTypes, scopeTypes } from './constants';
+export { getAttributesFromScopeId } from './getAttributesFromScopeId';